+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 
 - chore: add badges in `README.md` and classifiers in `pyproject.toml`
+- refactor: move the common code for manipulating the signature of the wrapped functions in `WithStore` and `Autorun` to a utility function
 
 ## Version 0.23.0